move core Profile logic from queryall-lib/ProfileUtils to queryall-api/ProfileMatch
